requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
groq>=0.4.2
requests>=2.31.0
pandas>=2.2.0
httpx>=0.24.1
pytz>=2024.1